LayoutTests/http/wpt/service-workers/fetch-service-worker-navigation-preload.https.html

<!doctype html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/routines.js"></script>
</head>
<body>
<script>
var activeWorker;
var frame;
const channel = new MessageChannel();
const uuid = token();
const url = "/WebKit/service-workers/resources/fetch-service-worker-preload-script.py?useNavigationPreloadPromise=true&token=" + uuid;

function waitUntilActivating()
{
    return new Promise(resolve => {
        channel.port2.onmessage = (event) => {
            if (event.data === "activating")
                resolve();
        };
    });
}

function triggerActivation()
{
    activeWorker.postMessage("activate");
}

promise_test(async (test) => {
    if (window.testRunner) {
        testRunner.setUseSeparateServiceWorkerProcess(true);
        await fetch("").then(() => { }, () => { });
    }

    let registration = await navigator.serviceWorker.register("/WebKit/service-workers/fetch-service-worker-preload-worker.js", { scope : url });
    if (!registration.installing) {
        registration.unregister();
        registration = await navigator.serviceWorker.register("/WebKit/service-workers/fetch-service-worker-preload-worker.js", { scope : url });
    }

    activeWorker = registration.installing;
    activeWorker.postMessage({ port: channel.port1 }, [channel.port1]);

    await waitUntilActivating();

    if (registration.navigationPreload)
        await registration.navigationPreload.enable();

    triggerActivation();
}, "Setup activating worker");

promise_test(async (test) => {
    await fetch(url + "&value=use-preload2", { method: 'POST' });


    const frame = await withIframe(url);
    assert_equals(frame.contentWindow.value, "use-preload2");

    // We should have only one GET fetch to url: the service worker preload
    const response = await fetch(url + "&count=True");
    assert_equals(await response.text(), "1");
}, "Service worker preloadResponse does not trigger an additional load");
</script>
</body>
</html>
